+++ test/integration/query.test.ts
@@ -1,18 +1,16 @@
-import { execSync } from 'child_process';
 import fetch from 'cross-fetch';
 import dotenv from 'dotenv';
 import { join } from 'path';
 import { afterAll, beforeAll, describe, expect, test } from 'vitest';
-import { BaseClient, contains, isXataRecord, lt, Repository, XataApiClient } from '../packages/client/src';
+import { BaseClient, contains, isXataRecord, lt, Repository, XataApiClient } from '../../packages/client/src';
 import {
   Paginable,
   PAGINATION_DEFAULT_SIZE,
   PAGINATION_MAX_OFFSET,
   PAGINATION_MAX_SIZE
-} from '../packages/client/src/schema/pagination';
-import { getCurrentBranchName } from '../packages/client/src/util/config';
-import { User, UserRecord, XataClient } from '../packages/codegen/example/xata';
-import { mockUsers, teamColumns, userColumns } from './mock_data';
+} from '../../packages/client/src/schema/pagination';
+import { User, UserRecord, XataClient } from '../../packages/codegen/example/xata';
+import { mockUsers, teamColumns, userColumns } from '../mock_data';
 
 // Get environment variables before reading them
 dotenv.config({ path: join(process.cwd(), '.envrc') });
@@ -30,7 +28,7 @@ const api = new XataApiClient({ apiKey, fetch });
 beforeAll(async () => {
   const id = Math.round(Math.random() * 100000);
 
-  const database = await api.databases.createDatabase(workspace, `sdk-integration-test-${id}`);
+  const database = await api.databases.createDatabase(workspace, `sdk-integration-test-query-${id}`);
   databaseName = database.databaseName;
 
   client = new XataClient({
@@ -563,371 +561,5 @@ describe('integration tests', () => {
 
     expect(team.owner?.id).toEqual(owner.id);
     expect(updated.owner?.id).toEqual(owner2.id);
-  });
-});
-
-describe('record creation', () => {
-  test('create single team without id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    expect(team.id).toBeDefined();
-    expect(team.name).toBe('Team ships');
-  });
-
-  test('create multiple teams without ids', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    expect(teams).toHaveLength(2);
-    expect(teams[0].id).toBeDefined();
-    expect(teams[0].name).toBe('Team cars');
-    expect(teams[0].read).toBeDefined();
-    expect(teams[1].id).toBeDefined();
-    expect(teams[1].name).toBe('Team planes');
-    expect(teams[1].read).toBeDefined();
-  });
-
-  test('create user with id', async () => {
-    const user = await client.db.users.create('a-unique-record-john-4', {
-      full_name: 'John Doe 4',
-      email: 'john4@doe.com'
-    });
-
-    const apiUser = await client.db.users.filter({ id: user.id }).getFirst();
-    if (!apiUser) throw new Error('No user found');
-
-    expect(user.id).toBe('a-unique-record-john-4');
-    expect(user.read).toBeDefined();
-    expect(user.full_name).toBe('John Doe 4');
-
-    expect(user.id).toBe(apiUser.id);
-    expect(user.full_name).toBe(apiUser.full_name);
-    expect(user.email).toBe(apiUser.email);
-
-    expect(
-      client.db.users.create('a-unique-record-john-4', {
-        full_name: 'John Doe 5',
-        email: 'john5@doe.com'
-      })
-    ).rejects.toHaveProperty('status', 422);
-  });
-
-  test('create user with inlined id', async () => {
-    const user = await client.db.users.create({
-      id: 'a-unique-record-john-5',
-      full_name: 'John Doe 5',
-      email: 'john5@doe.com'
-    });
-
-    const apiUser = await client.db.users.filter({ id: user.id }).getFirst();
-    if (!apiUser) throw new Error('No user found');
-
-    expect(user.id).toBe('a-unique-record-john-5');
-    expect(user.read).toBeDefined();
-    expect(user.full_name).toBe('John Doe 5');
-
-    expect(user.id).toBe(apiUser.id);
-    expect(user.full_name).toBe(apiUser.full_name);
-    expect(user.email).toBe(apiUser.email);
-  });
-
-  test('create user with empty id is not allowed', async () => {
-    expect(
-      client.db.users.create('', {
-        full_name: 'John Doe 3',
-        email: 'john3@doe.com'
-      })
-    ).rejects.toMatchInlineSnapshot(`[Error: The id can't be empty]`);
-  });
-
-  test('create user with empty inline id is not allowed', async () => {
-    expect(
-      client.db.users.create({
-        id: '',
-        full_name: 'John Doe 3',
-        email: 'john3@doe.com'
-      })
-    ).rejects.toMatchInlineSnapshot(`[Error: The id can't be empty]`);
-  });
-
-  test('create user with falsy id is not allowed', async () => {
-    expect(
-      //@ts-expect-error
-      client.db.users.create(null, {
-        full_name: 'John Doe 3',
-        email: 'john3@doe.com'
-      })
-    ).rejects.toMatchInlineSnapshot(`[Error: Invalid arguments for create method]`);
-  });
-
-  test("create multiple with empty array doesn't create anything", async () => {
-    const teams = await client.db.teams.create([]);
-    expect(teams).toHaveLength(0);
-  });
-});
-
-describe('record update', () => {
-  test('update single team', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    const updatedTeam = await client.db.teams.update(team.id, { name: 'Team boats' });
-
-    expect(updatedTeam.id).toBe(team.id);
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-    if (!apiTeam) throw new Error('No team found');
-
-    expect(updatedTeam.name).toBe('Team boats');
-    expect(apiTeam.name).toBe('Team boats');
-  });
-
-  test('update multiple teams', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    const updatedTeams = await client.db.teams.update(teams.map((team) => ({ ...team, name: 'Team boats' })));
-
-    expect(updatedTeams).toHaveLength(2);
-
-    const apiTeams = await client.db.teams.filter({ $any: teams.map((t) => ({ id: t.id })) }).getAll();
-
-    expect(apiTeams).toHaveLength(2);
-    expect(apiTeams[0].name).toBe('Team boats');
-    expect(apiTeams[1].name).toBe('Team boats');
-  });
-
-  test('update team with inline id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    const updatedTeam = await client.db.teams.update({ id: team.id, name: 'Team boats' });
-
-    expect(updatedTeam.id).toBe(team.id);
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-
-    expect(updatedTeam.name).toBe('Team boats');
-    expect(apiTeam?.name).toBe('Team boats');
-  });
-
-  test("update many with empty array doesn't update anything", async () => {
-    const updatedTeams = await client.db.teams.update([]);
-    expect(updatedTeams).toHaveLength(0);
-  });
-});
-
-describe('record deletion', () => {
-  test('delete single team with id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    await client.db.teams.delete(team.id);
-
-    const copy = await team.read();
-
-    expect(copy).toBeNull();
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-
-    expect(apiTeam).toBeNull();
-  });
-
-  test('delete multiple teams with id list', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    await client.db.teams.delete(teams.map((team) => team.id));
-
-    const apiTeams = await client.db.teams.filter({ $any: teams.map((t) => ({ id: t.id })) }).getAll();
-
-    expect(apiTeams).toHaveLength(0);
-  });
-
-  test('delete single team with id in object', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    await client.db.teams.delete(team);
-
-    const copy = await team.read();
-
-    expect(copy).toBeNull();
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-
-    expect(apiTeam).toBeNull();
-  });
-
-  test('delete multiple teams with id in object', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    await client.db.teams.delete(teams);
-
-    const apiTeams = await client.db.teams.filter({ $any: teams.map((t) => ({ id: t.id })) }).getAll();
-
-    expect(apiTeams).toHaveLength(0);
-  });
-
-  test('delete team with own operation', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    await team.delete();
-
-    const copy = await team.read();
-
-    expect(copy).toBeNull();
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-
-    expect(apiTeam).toBeNull();
-  });
-});
-
-describe('record create or update', () => {
-  test('create or update single team with id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    const updatedTeam = await client.db.teams.createOrUpdate(team.id, { name: 'Team boats' });
-
-    expect(updatedTeam.id).toBe(team.id);
-    expect(updatedTeam.read).toBeDefined();
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-
-    expect(updatedTeam.name).toBe('Team boats');
-    expect(apiTeam?.name).toBe('Team boats');
-  });
-
-  test('create or update team with inline id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    const updatedTeam = await client.db.teams.createOrUpdate({ id: team.id, name: 'Team boats' });
-
-    expect(updatedTeam.id).toBe(team.id);
-    expect(updatedTeam.read).toBeDefined();
-
-    const apiTeam = await client.db.teams.filter({ id: team.id }).getFirst();
-
-    expect(updatedTeam.name).toBe('Team boats');
-    expect(apiTeam?.name).toBe('Team boats');
-  });
-
-  test('create or update multiple teams', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    const updatedTeams = await client.db.teams.createOrUpdate(
-      teams.map((team) => ({ id: team.id, name: 'Team boats' }))
-    );
-
-    expect(updatedTeams).toHaveLength(2);
-    expect(updatedTeams[0].read).toBeDefined();
-
-    const apiTeams = await client.db.teams.filter({ $any: teams.map((t) => ({ id: t.id })) }).getAll();
-
-    expect(apiTeams).toHaveLength(2);
-
-    expect(apiTeams[0].name).toBe('Team boats');
-    expect(apiTeams[1].name).toBe('Team boats');
-  });
-
-  test("create or update many with empty array doesn't create or update anything", async () => {
-    const updatedTeams = await client.db.teams.createOrUpdate([]);
-    expect(updatedTeams).toHaveLength(0);
-  });
-});
-
-describe('record read', () => {
-  test('read single team with id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
-
-    const copy = await client.db.teams.read(team.id);
-    expect(copy).toBeDefined();
-    expect(copy?.id).toBe(team.id);
-  });
-
-  test('read multiple teams with id list', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    const copies = await client.db.teams.read(teams.map((team) => team.id));
-    expect(copies).toHaveLength(2);
-    expect(copies[0]?.id).toBe(teams[0].id);
-    expect(copies[1]?.id).toBe(teams[1].id);
-  });
-
-  test("read single and return null if team doesn't exist", async () => {
-    const copy = await client.db.teams.read('does-not-exist');
-    expect(copy).toBeNull();
-  });
-
-  test("read multiple teams with id list and ignores a team if doesn't exist", async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
-
-    const copies = await client.db.teams.read(teams.map((team) => team.id).concat(['does-not-exist']));
-    expect(copies).toHaveLength(2);
-    expect(copies[0]?.id).toBe(teams[0].id);
-    expect(copies[1]?.id).toBe(teams[1].id);
-  });
-
-  test('read multiple with empty array', async () => {
-    const copies = await client.db.teams.read([]);
-    expect(copies).toHaveLength(0);
-  });
-
-  test('read multiple with falsy values, throws', async () => {
-    // @ts-ignore
-    expect(client.db.teams.read([null, undefined, false, 0, ''])).rejects.toThrow();
-  });
-});
-
-describe('getBranch', () => {
-  const envValues = { ...process.env };
-  const gitBranch = execSync('git branch --show-current', { encoding: 'utf-8' }).trim();
-
-  afterAll(() => {
-    // Reset env variable values
-    process.env = envValues;
-  });
-
-  test('uses an env variable if it is set', async () => {
-    const branchName = 'using-env-variable';
-
-    await api.branches.createBranch(workspace, databaseName, branchName);
-
-    const getBranchOptions = {
-      apiKey,
-      databaseURL: `https://${workspace}.xata.sh/db/${databaseName}`,
-      fetchImpl: fetch
-    };
-
-    process.env = { NODE_ENV: 'development', XATA_BRANCH: branchName };
-    expect(await getCurrentBranchName(getBranchOptions)).toEqual(branchName);
-
-    process.env = { NODE_ENV: 'development', VERCEL_GIT_COMMIT_REF: branchName };
-    expect(await getCurrentBranchName(getBranchOptions)).toEqual(branchName);
-
-    process.env = { NODE_ENV: 'development', CF_PAGES_BRANCH: branchName };
-    expect(await getCurrentBranchName(getBranchOptions)).toEqual(branchName);
-
-    process.env = { NODE_ENV: 'development', BRANCH: branchName };
-    expect(await getCurrentBranchName(getBranchOptions)).toEqual(branchName);
-  });
-
-  test('uses `main` if no env variable is set is not set and there is not associated git branch', async () => {
-    process.env = { NODE_ENV: 'development' };
-    const branch = await getCurrentBranchName({
-      apiKey,
-      databaseURL: `https://${workspace}.xata.sh/db/${databaseName}`,
-      fetchImpl: fetch
-    });
-
-    expect(branch).toEqual('main');
-  });
-
-  test('uses the git branch name if branch exists', async () => {
-    process.env = { NODE_ENV: 'development' };
-
-    await api.branches.createBranch(workspace, databaseName, gitBranch);
-
-    const branch = await getCurrentBranchName({
-      apiKey,
-      databaseURL: `https://${workspace}.xata.sh/db/${databaseName}`,
-      fetchImpl: fetch
-    });
-
-    expect(branch).toEqual(gitBranch);
   });
 });
